requirements.txt
Django==5.0.2
djangorestframework==3.15.2
sqlparse==0.4.4
typing_extensions==4.9.0
PyJWT==1.7.1
asgiref==3.8.1
gunicorn==23.0.0
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
tzdata==2024.2